Изменил тип данных и добавил мысли о проекте

+++ dags/sql/creating_schema.sql
@@ -5,7 +5,7 @@ CREATE SCHEMA IF NOT EXISTS stg;
 CREATE TABLE IF NOT EXISTS stg.adv (
 daily_time_spent_on_site float,
 age SMALLINT,
-area_income float,
+area_income NUMERIC(7, 2),
 daily_internet_usage float,
 ad_topic_line VARCHAR(55),
 city VARCHAR(30),
